# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ISU Mission Control API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface IssPosition {
  latitude: number;
  longitude: number;
  altitude: number;
  velocity: number;
  timestamp: number;
  visibility: string;
}

export interface IssPass {
  risetime: number;
  duration: number;
  maxElevation: number;
}

export type IssPassesResponseLocation = {
  latitude: number;
  longitude: number;
  name: string;
};

export interface IssPassesResponse {
  passes: IssPass[];
  location: IssPassesResponseLocation;
}

export interface CrewMember {
  name: string;
  craft: string;
}

export interface IssCrewResponse {
  people: CrewMember[];
  number: number;
}

export interface IssSummary {
  position: IssPosition;
  nextPass?: IssPass;
  crewCount: number;
  missionDay: number;
  orbitsCompleted: number;
}

export interface Conversation {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
}

export interface Message {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface ConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
  messages: Message[];
}

export interface CreateConversationBody {
  title: string;
}

export interface SendMessageBody {
  content: string;
}

export interface ErrorResponse {
  error: string;
}