# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ISU Mission Control API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get current ISS position, altitude, and velocity
 */
export const GetIssPositionResponse = zod.object({
  latitude: zod.number(),
  longitude: zod.number(),
  altitude: zod.number(),
  velocity: zod.number(),
  timestamp: zod.number(),
  visibility: zod.string(),
});

/**
 * @summary Get upcoming ISS passes over ISU campus in Strasbourg
 */
export const GetIssPassesResponse = zod.object({
  passes: zod.array(
    zod.object({
      risetime: zod.number(),
      duration: zod.number(),
      maxElevation: zod.number(),
    }),
  ),
  location: zod.object({
    latitude: zod.number(),
    longitude: zod.number(),
    name: zod.string(),
  }),
});

/**
 * @summary Get current ISS crew members
 */
export const GetIssCrewResponse = zod.object({
  people: zod.array(
    zod.object({
      name: zod.string(),
      craft: zod.string(),
    }),
  ),
  number: zod.number(),
});

/**
 * @summary Get aggregated ISS telemetry snapshot for dashboard
 */
export const GetIssSummaryResponse = zod.object({
  position: zod.object({
    latitude: zod.number(),
    longitude: zod.number(),
    altitude: zod.number(),
    velocity: zod.number(),
    timestamp: zod.number(),
    visibility: zod.string(),
  }),
  nextPass: zod
    .object({
      risetime: zod.number(),
      duration: zod.number(),
      maxElevation: zod.number(),
    })
    .optional(),
  crewCount: zod.number(),
  missionDay: zod.number(),
  orbitsCompleted: zod.number(),
});

/**
 * @summary List all AI mentor conversations
 */
export const ListConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListConversationsResponse = zod.array(
  ListConversationsResponseItem,
);

/**
 * @summary Create a new AI mentor conversation
 */
export const CreateConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get a conversation with its messages
 */
export const GetConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.string(),
    }),
  ),
});

/**
 * @summary Send a message to AI mentor (SSE streaming)
 */
export const SendMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendMessageBody = zod.object({
  content: zod.string(),
});